implements form component, creates edit page (update function not working yet)
